fix type cast
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@31458 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSString.m
@@ -300,7 +300,7 @@ setup(BOOL rerun)
       equalImp = (BOOL (*)(id, SEL, id))
 	[NSStringClass instanceMethodForSelector: equalSel];
       hashSel = @selector(hash);
-      hashImp = (unsigned (*)(id, SEL))
+      hashImp = (NSUInteger (*)(id, SEL))
 	[GSStringClass instanceMethodForSelector: hashSel];
 
       caiSel = @selector(characterAtIndex:);
